+++ examples/sdk-app-example/app/src/main/java/com/example/bugsnag/android/BaseCrashyActivity.kt
@@ -11,7 +11,9 @@ import com.bugsnag.android.Configuration
 import com.bugsnag.android.Severity
 import com.example.foo.CrashyClass
 import com.google.android.material.snackbar.Snackbar
+import java.util.Date
 import java.util.HashMap
+
 
 open class BaseCrashyActivity : AppCompatActivity() {
 
@@ -50,6 +52,20 @@ open class BaseCrashyActivity : AppCompatActivity() {
             showSnackbar()
         }
     }
+
+    override fun onResume() {
+        val members = listOf(
+            mapOf("Group Members 1" to "Adam"),
+            mapOf("Group Members 2" to "Alice")
+        )
+        val lastResumeTime = mapOf("Last Resume Time" to Date())
+
+        Bugsnag.addMetadata("Custom Data", "members", members)
+        Bugsnag.addMetadata("Last Resume Time", "Member Last Resume Time", lastResumeTime)
+
+        super.onResume()
+    }
+
 
     /**
      * Throws an unhandled Exception. Bugsnag will automatically capture any uncaught exceptions